requirements.txt
aiohttp==3.11.2
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
distro==1.9.0
frozenlist==1.5.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.7
langchain-core>=0.2.38,<0.4
langchain-openai==0.1.15
langchain-text-splitters==0.2.2
langchainhub==0.1.20
langgraph==0.1.7
langsmith==0.1.85
multidict==6.1.0
numpy==1.26.4
openai==1.54.4
orjson==3.10.6
packaging==24.2
pillow==10.4.0
playwright==1.45.0
psutil==6.1.0
pydantic==2.9.2
pydantic_core==2.23.4
pyee==11.1.0
PyYAML==6.0.2
python-dotenv
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.31
tenacity==8.5.0
tiktoken==0.7.0
tqdm==4.67.0
types-requests==2.32.0.20240622
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.17.1
